% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with the columns v1, v2 and w
v1 and v2 are the nodes and w is the weight between them}

\item{init_node}{The node where the algorithm starts}
}
\value{
A vector with the shortest distance from init_node
to all nodes in the graph
}
\description{
Finds the shortest distance from one node to all other nodes
The function starts at the chosen node
}
\details{
It checks the connected nodes and updates the shortest distance
This continues until all nodes have been visited
}
\references{
Wikipedia: Dijkstra's algorithm
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
