% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph}
\format{
Includes a data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Starting node.}
  \item{v2}{End node.}
  \item{w}{Distance between starting and end nodes.}
}
}
\usage{
wiki_graph
}
\description{
This is the graph used in Dijkstra's Algorithm. It comes from Wikipedia.
}
\references{
Wikipedia: Dijkstra's algorithm
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
